+++ app/assets/javascripts/angular/controllers/UserCrtl.js.coffee
@@ -39,10 +39,11 @@
 	    	$scope.status = "There was an error deleting this user: " + error.message
 
 	$scope.deleteUser = ->
-		User.deleteUser($scope.id).success((data) ->
-	    	$location.path('/admin/users/')
-	  	).error (error) ->
-	    	$scope.status = "There was an error deleting this user: " + error.message
+		if (confirm("Are you sure you want to delete this user?") is true)
+			User.deleteUser($scope.id).success((data) ->
+		    	$location.path('/admin/users/')
+		  	).error (error) ->
+		    	$scope.status = "There was an error deleting this user: " + error.message
 
 	$scope.getNextOfKins = ->
 		if $scope.allowed(3)
@@ -52,10 +53,11 @@
 		    	$scope.status = "Unable to load user data: " + error.message
 
 	$scope.deleteNextOfKin = (id, index) ->
-		NextOfKin.deleteNextOfKin($scope.id, id).success((data) ->
-			$scope.nextOfKins.splice(index, 1)
-		).error (error) ->
-			$scope.status = "There was an error deleting your data " + error.message
+		if (confirm("Are you sure you want to delete this next of kin?") is true)
+			NextOfKin.deleteNextOfKin($scope.id, id).success((data) ->
+				$scope.nextOfKins.splice(index, 1)
+			).error (error) ->
+				$scope.status = "There was an error deleting your data " + error.message
 
 	$scope.getDisciplinaryCards = ->
 		User.getDisciplinaryCards($scope.id).success((data) ->
@@ -70,10 +72,11 @@
 	    	$scope.error = errors.errors
 
 	$scope.deleteDisciplinaryCard = (id, index) ->
-		DisciplinaryCard.deleteDisciplinaryCard(id).success((data) ->
-			$scope.disciplinaryCards.splice(index, 1)
-		).error (error) ->
-	    	$scope.status = "Unable to load user data: " + error.message
+		if (confirm("Are you sure you want to delete this disciplinary card?") is true)
+			DisciplinaryCard.deleteDisciplinaryCard(id).success((data) ->
+				$scope.disciplinaryCards.splice(index, 1)
+			).error (error) ->
+		    	$scope.status = "Unable to load user data: " + error.message
 
 	$scope.yesOrNo = (bool) ->
 		if bool == "true"
